PngViewer: Rewrote the pixel by pixel SDL rendering
This doesn't use SDL_gfx anymore (it's pixelRGBA function) but rather uses a pointer to the pixels of the main surface of the window.
It then goes through each pixel and uses SDL_MapRGBA to get the pixel data to directly inserts it.
This greatly improves performance and memory usage when drawing the image to the screen

(Also a small patch on the parser to free a unmanaged allocation)